Update location data types in migration files
Correctly set Latitude and Longitude properties to double in migration files to align with entity definitions and prevent phantom migrations.

+++ apps/services/careconnect/CareConnect.Infrastructure/Data/Migrations/20260328190001_AddProviderGeoLocation.cs
@@ -11,13 +11,13 @@ namespace CareConnect.Infrastructure.Data.Migrations
         /// <inheritdoc />
         protected override void Up(MigrationBuilder migrationBuilder)
         {
-            migrationBuilder.AddColumn<decimal>(
+            migrationBuilder.AddColumn<double>(
                 name: "Latitude",
                 table: "Providers",
                 type: "decimal(10,7)",
                 nullable: true);
 
-            migrationBuilder.AddColumn<decimal>(
+            migrationBuilder.AddColumn<double>(
                 name: "Longitude",
                 table: "Providers",
                 type: "decimal(10,7)",
